+++ Directory.Build.props
@@ -9,7 +9,7 @@
 			<IncludeAssets>runtime; build; native; contentfiles; analyzers</IncludeAssets>
 		</PackageReference>
 		<!-- Microsoft.VisualStudio.Threading.Analyzers has the async void analyzers. -->
-		<PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.12.19" PrivateAssets="all" />
+		<PackageReference Include="Microsoft.VisualStudio.Threading.Analyzers" Version="17.13.2" PrivateAssets="all" />
 		<PackageReference Include="GooseAnalyzers" Version="1.0.0" PrivateAssets="all" />
 	</ItemGroup>
