+++ src/ApplicationTemplate.sln
@@ -11,24 +11,34 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "root", "root", "{8711E408-6
 		..\.azure-pipelines.yml = ..\.azure-pipelines.yml
 		..\.editorconfig = ..\.editorconfig
 		..\.gitignore = ..\.gitignore
+		..\APP_README.md = ..\APP_README.md
+		..\CODE_OF_CONDUCT.md = ..\CODE_OF_CONDUCT.md
+		..\CONTRIBUTING.md = ..\CONTRIBUTING.md
 		..\Directory.Build.props = ..\Directory.Build.props
 		..\nuget.config = ..\nuget.config
+		..\README.md = ..\README.md
 		..\stylecop.json = ..\stylecop.json
 	EndProjectSection
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "doc", "doc", "{4A31B4ED-94EB-43BC-91DD-7B142DEB887C}"
 	ProjectSection(SolutionItems) = preProject
 		..\doc\Architecture.md = ..\doc\Architecture.md
+		..\doc\AzurePipelines.md = ..\doc\AzurePipelines.md
 		..\CHANGELOG.md = ..\CHANGELOG.md
 		..\doc\Configuration.md = ..\doc\Configuration.md
+		..\doc\DataLoading.md = ..\doc\DataLoading.md
+		..\doc\DefaultAnalytics.md = ..\doc\DefaultAnalytics.md
 		..\doc\DependencyInjection.md = ..\doc\DependencyInjection.md
+		..\doc\Diagnostics.md = ..\doc\Diagnostics.md
 		..\doc\Environments.md = ..\doc\Environments.md
 		..\doc\ErrorHandling.md = ..\doc\ErrorHandling.md
+		..\doc\ForcedUpdate.md = ..\doc\ForcedUpdate.md
 		..\doc\HTTP.md = ..\doc\HTTP.md
+		..\doc\KillSwitch.md = ..\doc\KillSwitch.md
 		..\doc\Localization.md = ..\doc\Localization.md
 		..\doc\Logging.md = ..\doc\Logging.md
 		..\doc\PlatformSpecifics.md = ..\doc\PlatformSpecifics.md
-		..\README.md = ..\README.md
+		..\doc\Reviews.md = ..\doc\Reviews.md
 		..\doc\Scheduling.md = ..\doc\Scheduling.md
 		..\doc\Serialization.md = ..\doc\Serialization.md
 		..\doc\Startup.md = ..\doc\Startup.md
@@ -64,8 +74,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "build", "build", "{6E8378CB
 		..\build\canary-merge.yml = ..\build\canary-merge.yml
 		..\build\gitversion-config.yml = ..\build\gitversion-config.yml
 		..\build\stage-build.yml = ..\build\stage-build.yml
-		..\build\stage-release-firebase-app-distribution.yml = ..\build\stage-release-firebase-app-distribution.yml
 		..\build\stage-release-appstore.yml = ..\build\stage-release-appstore.yml
+		..\build\stage-release-firebase-app-distribution.yml = ..\build\stage-release-firebase-app-distribution.yml
 		..\build\stage-release-googleplay.yml = ..\build\stage-release-googleplay.yml
 		..\build\steps-build-android.yml = ..\build\steps-build-android.yml
 		..\build\steps-build-ios.yml = ..\build\steps-build-ios.yml
