+++ src/app/ApplicationTemplate.Mobile/ApplicationTemplate.Mobile.csproj
@@ -22,23 +22,26 @@
 		<PackageReference Include="Chinook.SectionsNavigation.Uno.WinUI" Version="3.0.2" />
 		<PackageReference Include="ExtendedSplashScreen.Uno.WinUI" Version="2.0.0" />
 		<PackageReference Include="MessageDialogService.Uno.WinUI" Version="2.0.0" />
-		<PackageReference Include="Microsoft.Maui.Essentials" Version="9.0.30" />
+		<PackageReference Include="Microsoft.Maui.Essentials" Version="9.0.50" />
 		<PackageReference Include="Nventive.Persistence.Uno.WinUI" Version="0.5.0" />
 		<PackageReference Include="Nventive.View.Uno.WinUI" Version="0.6.0" />
 		<PackageReference Include="Reactive.Annex.Uno.WinUI" Version="2.0.0" />
 		<PackageReference Include="ReviewService.NativePrompters" Version="1.0.0" />
 		<PackageReference Include="Uno.CommunityToolkit.WinUI" Version="7.1.204" />
-		<PackageReference Include="Uno.Material.WinUI" Version="5.3.1" />
+		<PackageReference Include="Uno.Material.WinUI" Version="5.4.0" />
 		<PackageReference Include="Uno.Microsoft.Xaml.Behaviors.Interactivity.WinUI" Version="2.4.2" />
 		<PackageReference Include="Uno.Microsoft.Xaml.Behaviors.WinUI.Managed" Version="2.4.2" />
-		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="6.3.7" />
-		<PackageReference Include="Uno.WinUI" Version="5.6.30" />
-		<PackageReference Include="Uno.WinUI.DevServer" Version="5.6.30" Condition="'$(Configuration)'=='Debug'" />
-		<PackageReference Include="Uno.WinUI.Lottie" Version="5.6.30" />
-		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="5.6.30" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.1" />
+		<PackageReference Include="Uno.Toolkit.WinUI.Material" Version="6.4.3" />
+		<PackageReference Include="Uno.WinUI" Version="5.6.81" />
+		<PackageReference Include="Uno.WinUI.DevServer" Version="5.6.81" Condition="'$(Configuration)'=='Debug'" />
+		<PackageReference Include="Uno.WinUI.Lottie" Version="5.6.81" />
+		<PackageReference Include="Uno.UI.Adapter.Microsoft.Extensions.Logging" Version="5.6.81" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Console" Version="9.0.3" />
 		<PackageReference Include="Serilog.Sinks.Xamarin" Version="1.0.0" />
 		<PackageReference Include="Serilog.Sinks.File" Version="6.0.0" />
+
+		<!-- It would normally be a transitive dependency, however, the version included with one of the packages has a vulnerability. -->
+		<PackageReference Include="System.Private.Uri" Version="4.3.2" />
 	</ItemGroup>
 	<ItemGroup>
 		<ProjectReference Include="..\ApplicationTemplate.Access\ApplicationTemplate.Access.csproj" />
@@ -102,6 +105,9 @@
 				<PackageReference Include="Uno.UniversalImageLoader" Version="1.9.37" />
 				<PackageReference Include="Mono.AotProfiler.Android" Version="7.0.0" />
 				<PackageReference Include="Xamarin.Kotlin.StdLib.Jdk8" Version="2.0.21.2" />
+
+				<!-- Required with the last version of Uno otherwise an exception is thrown in the main activity. -->
+				<PackageReference Include="Xamarin.AndroidX.AppCompat" Version="1.7.0.5" />
 			</ItemGroup>
 			<ItemGroup>
 				<AndroidAotProfile Include="custom.aprof" />
