+++ src/app/ApplicationTemplate.Windows/Properties/PublishProfiles/win-arm64.pubxml
@@ -5,16 +5,10 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
 <Project ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <PropertyGroup>
     <PublishProtocol>FileSystem</PublishProtocol>
-    <Platform>arm64</Platform>
+    <Platform>ARM64</Platform>
     <RuntimeIdentifier>win-arm64</RuntimeIdentifier>
     <PublishDir>bin\$(Configuration)\$(TargetFramework)\$(RuntimeIdentifier)\publish\</PublishDir>
     <SelfContained>true</SelfContained>
     <PublishSingleFile>False</PublishSingleFile>
-    <PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
-    <PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-    <!--
-    See https://github.com/microsoft/CsWinRT/issues/373
-    <PublishTrimmed>True</PublishTrimmed>
-    -->
   </PropertyGroup>
 </Project>
