+++ src/app/ApplicationTemplate.Windows/Properties/PublishProfiles/win-x64.pubxml
@@ -10,11 +10,5 @@ https://go.microsoft.com/fwlink/?LinkID=208121.
     <PublishDir>bin\$(Configuration)\$(TargetFramework)\$(RuntimeIdentifier)\publish\</PublishDir>
     <SelfContained>true</SelfContained>
     <PublishSingleFile>False</PublishSingleFile>
-    <PublishReadyToRun Condition="'$(Configuration)' == 'Debug'">False</PublishReadyToRun>
-    <PublishReadyToRun Condition="'$(Configuration)' != 'Debug'">True</PublishReadyToRun>
-    <!--
-    See https://github.com/microsoft/CsWinRT/issues/373
-    <PublishTrimmed>True</PublishTrimmed>
-    -->
   </PropertyGroup>
 </Project>
